requirements.txt
aiogram==2.20
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
certifi==2021.10.8
charset-normalizer==2.0.12
environs==9.5.0
frozenlist==1.3.0
idna==3.3
marshmallow==3.15.0
multidict==6.0.2
packaging==21.3
psycopg2-binary==2.9.3
pyparsing==3.0.8
python-dotenv==0.20.0
pytz==2022.1
yarl==1.7.2
